moving separate_dot_field_hierarchy
