+++ nexus_constructor/json/filewriter_json_writer.py
@@ -2,13 +2,10 @@ import h5py
 import numpy as np
 import uuid
 import logging
-from typing import Union, Dict
+from typing import Union, Dict, Any, List, Tuple
 
 from nexus_constructor.instrument import Instrument
-from nexus_constructor.json.helpers import (
-    object_to_json_file,
-    _separate_dot_field_group_hierarchy,
-)
+from nexus_constructor.json.helpers import object_to_json_file
 from nexus_constructor.nexus.nexus_wrapper import get_nx_class, get_name_of_node
 
 NexusObject = Union[h5py.Group, h5py.Dataset, h5py.SoftLink]
@@ -273,3 +270,19 @@ def create_writer_commands(
         stop_cmd["service_id"] = service_id
 
     return write_cmd, stop_cmd
+
+
+def _separate_dot_field_group_hierarchy(
+    item_dict: Dict[Any, Any],
+    dots_in_field_name: List[str],
+    item: Tuple[str, h5py.Group],
+):
+    previous_group = item_dict
+    for subgroup in dots_in_field_name:
+        # do not overwrite a group unless it doesn't yet exist
+        if subgroup not in previous_group:
+            previous_group[subgroup] = dict()
+        if subgroup == dots_in_field_name[-1]:
+            # set the value of the field to the last item in the list
+            previous_group[subgroup] = item[...][()]
+        previous_group = previous_group[subgroup]
